test(rename): move rename tests to rename_tests.rs
Same reason as the earlier hover test move: inline unwrap()/expect() in
test code tripped the pre-push guideline check, which only recognizes test
code by file path (tests.rs / a tests/ dir), not #[cfg(test)] blocks inline
in a non-test-suffixed file. Moving the whole module (not just the new
tests) matches every other feature's convention in this codebase.
